Beta 1 of Version 1.7.0
This new minor release doesn't bring a lot but it has a long awaited change to sorting and other click-intense (for the mod) tasks.

### The story so far
The current approach was to always wait for the server to confirm that the interaction was successful. This is a really robust system even on slow servers but can cause serious slowness or lags on slow servers.

### What changed?

In the 1.17 snapshots Mojang removed the confirmation events that this system was built around. This forced me to now switch to a time based system.

You can now configure the time in milliseconds between each packet getting sent. This means that you can clock up the sorting to really high speeds. Just keep in mind that servers might reject this or kick you for sending too many packets.

Since I didn't really test all functionality to work with this system this'll start as a beta but probably quickly stabilize.

### Other things
The config screen can now be opened through the keybinding again.

### PS
This mod is meant to be used in Vanilla and modded scenarios where everyone agrees upon its use. I do not endorse usage of this mod for cheating.