Version 1.10.3
It has been a long wait, but I'm back with a relatively big minor release :)

- Loyalty tridents will no longer be refilled.
- Single items will now be transported by just simulating a simple shift+click. This not only fixes bundles, but also increases performance quite a bit.
- Performance improvements for sorting.
- Sorting the inventory will no longer in any circumstance affect the hotbar.
- You may now open the config screen from the title screen via the keybinding.
- Fixed a bunch of crashes related to the config screen.
- Fixed another bunch of random crashes.

### 1.19.3 Support!

With 1.19.3, Mojang improved the creative inventory a lot.
Mouse Wheelie now offers a new sort mode based on the sorting of the creative inventory search tab.
Since the old default (raw id) basically was the creative inventory search before, the new creative sort mode is the new default.
Your config will be updated to reflect that, but you can still go back to the old inferior sorting :)

There's also a new config option to disable some optimizations with this new sort mode in favor of saving a bit of memory.